Fix observing being stopped whenever sell text changed to "ASK"
The observer was only handlng if the characterdata was updated, however whenever the sell text changed to "ASK"
a new <div> was inserted into the document. This caused the observer to just crap its pants because the span
it was looking at POOF suddenly didn't exist. To solve this the observer now checks the parent container of the sell
text and checks if either the char data was changed, OR an element was added/removed.

Some validation was also added (e.g. if the old value of the text is the same as the current one)
so it doesn't check everything at all times