Add hack to stop service worker from going to sleep
So service workers have this lovely feature that if they don't receive any messages
for 30 seconds, they go inactive. In theory a new message should wake the service
worker. In practice it doesn't until the page is refreshed and just keeps throwing the error
"trying to connect on a disconnected port". What's more after 5 minutes a service worker
will be terminated so it doesn't take up memory from the browser which is not exactly what
I want as it's constantly monitoring a site (for usually at least 90 minutes).
So to fix this a hack has been added, since communicating with an offscreen document
seems to keep the service worker alive.